1. Upgrade java Compatibility to JDK11. 2. Upgrade kotlin to v1.5.31. 3. Upgrade framework utils. 4. Upgrade gradle to v7.2, gradle plugin to v7.0.2 and adopt sources to new gradle. 5. Improve framework sources to new Kotlin.

+++ androidbase-sdk/src/main/java/com/leovp/androidbase/exts/kotlin/ByteArrayExt.kt
@@ -91,7 +91,7 @@ fun Long.toBytesLE(): ByteArray = ByteArray(Long.SIZE_BYTES).also { for (i in it
 // ==============================================
 
 fun Byte.toHexString(addPadding: Boolean = false) = let { if (addPadding) "%02X".format(it) else "%X".format(it) }
-fun ByteArray.toAsciiString(delimiter: CharSequence = ",") = map { it.toChar() }.joinToString(delimiter)
+fun ByteArray.toAsciiString(delimiter: CharSequence = ",") = map { it.toInt().toChar() }.joinToString(delimiter)
 
 /**
  * Attention.
